full_interaction_matrix
The code takes a numpy array with numeric values as an input and adds new columns, so that the output contains original columns plus the full set of interaction columns. I call 'interaction' as a product of two (or more) columns. For example, if the input contains columns a,b,c, then the output would contain columns: a, a*b, b, a*c, a*b*c, b*c, c.  The implementation is similar to the logic described in 'SAS User Guide, Specification of Effects' : 
https://support.sas.com/documentation/cdl/en/statug/63033/HTML/default/viewer.htm#statug_glm_sect027.htm

The function returns two object: list of labels and numpy array with all interaction terms.